+++ src/kernels.cuh
@@ -1,5 +1,6 @@
 #pragma once
 #include <cuda_runtime.h>
+#include <cuda_fp16.h>
 #include <stdio.h>
 
 enum WeightPrecision : int;
@@ -28,8 +29,8 @@ void cuda_rope(float* d_q, float* d_k, int pos, int dim, int kv_dim, int head_si
 void cuda_swiglu(float* d_hb, float* d_h1, float* d_h3, int size);
 void cuda_softmax_multihead(float* d_out, float* d_x, int n_heads, int seq_len, int att_stride);
 void cuda_scale_multihead(float* d_att, float scale, int n_heads, int seq_len, int att_stride);
-void cuda_aggregation_multihead(float* d_out, const float* d_v, const float* d_att, int n_heads, int seq_len, int head_size, int gqa_factor, int att_stride);
+void cuda_aggregation_multihead(float* d_out, const __half* d_v, const float* d_att, int n_heads, int seq_len, int head_size, int gqa_factor, int att_stride);
 void cuda_residual_add(float* d_out, const float* d_a, const float* d_b, int n);
-void cuda_scatter_kv(float* d_key_cache, float* d_value_cache, const float* d_k, const float* d_v, int layer, int pos, int n_kv_heads, int head_size, int seq_len);
-void cuda_multihead_gemv(float* d_out, float* d_q, float* d_k_cache, int layer, int pos, int n_heads, int n_kv_heads, int head_size, int seq_len);
+void cuda_scatter_kv(__half* d_key_cache, __half* d_value_cache, const float* d_k, const float* d_v, int layer, int pos, int n_kv_heads, int head_size, int seq_len);
+void cuda_multihead_gemv(float* d_out, float* d_q, __half* d_k_cache, int layer, int pos, int n_heads, int n_kv_heads, int head_size, int seq_len);
 void cuda_sample(float* d_logits, int vocab_size,float temperature,float topp,float penalty,int* d_history,int history_len,int* d_sampled_token,unsigned int* d_rng_state);
